reorganize addons to make force flag more generic
