+++ pkg/minikube/node/start.go
@@ -37,7 +37,6 @@ import (
 	kconst "k8s.io/kubernetes/cmd/kubeadm/app/constants"
 	cmdcfg "k8s.io/minikube/cmd/minikube/cmd/config"
 	"k8s.io/minikube/pkg/addons"
-	"k8s.io/minikube/pkg/addons/gcpauth"
 	"k8s.io/minikube/pkg/drivers/kic/oci"
 	"k8s.io/minikube/pkg/kapi"
 	"k8s.io/minikube/pkg/minikube/bootstrapper"
@@ -166,7 +165,7 @@ func Start(starter Starter, apiServer bool) (*kubeconfig.Settings, error) {
 	// enable addons, both old and new!
 	if starter.ExistingAddons != nil {
 		if viper.GetBool("force") {
-			gcpauth.Force = true
+			addons.Force = true
 		}
 		wg.Add(1)
 		go addons.Start(&wg, starter.Cfg, starter.ExistingAddons, config.AddonList)
