+++ site/content/en/docs/commands/addons.md
@@ -128,7 +128,7 @@ minikube addons enable dashboard
 ### Options
 
 ```
-      --force               If true, will force gcp-auth addon to be enabled on GCE.
+      --force               If true, will perform potentially dangerous operations. Use with discretion.
       --images string       Images used by this addon. Separated by commas.
       --registries string   Registries used by this addon. Separated by commas.
 ```
